🚚 Rename package `tests.fixtures.{repositories => packages}`
