+++ tests/fixtures/packages/adapters/__init__.py
