+++ tests/fixtures/packages/adapters/mercurial.py
