+++ tests/fixtures/packages/domain/stores.py
